requirements.txt
et_xmlfile==2.0.0
greenlet==3.3.2
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
